closer
letters are spaced and three of them are in english

+++ src/TextRenderer.h
@@ -11,7 +11,7 @@ struct Character {
     GLuint TextureID;   // ID handle of the glyph texture
     glm::ivec2 Size;    // Size of glyph
     glm::ivec2 Bearing; // Offset from baseline to left/top of glyph
-    int Advance;     // Offset to advance to next glyph
+    float Advance;     // Offset to advance to next glyph
 };
 
 class TextRenderer {
